QString: prevent crash in remove(pos, len)
If pos < 0, the code assumes it should count from end of string, i.e.
 pos += size(); if pos is still < 0, do nothing.

+++ src/corelib/text/qstring.cpp
@@ -3290,7 +3290,7 @@ QString &QString::remove(qsizetype pos, qsizetype len)
     if (pos < 0)  // count from end of string
         pos += size();
 
-    if (pos >= size() || len <= 0)
+    if (size_t(pos) >= size_t(size()) || len <= 0)
         return *this;
 
     len = std::min(len, size() - pos);
